+++ include/Pacman.hpp
@@ -20,7 +20,7 @@ class Pacman : public arc::IGame {
 		const Specs &getSpecs() const noexcept final;
 		bool processInteraction(Interaction &) noexcept final;
 		void envUpdate() noexcept final;
-		int getScore() noexcept final;
+		size_t getScore() noexcept final;
 	private:
 		Item &getItemFromName(const std::string &);
 		bool removeItem(const std::string &);
@@ -44,7 +44,7 @@ class Pacman : public arc::IGame {
 		Specs _spec;
 		Interaction _direction;
 		float _eating;
-		int _score;
+		size_t _score;
 };
 }
